requirements.txt
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2

fastapi==0.104.1
uvicorn[standard]==0.24.0

pydantic==2.5.0

pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0

flake8==6.1.0
black==23.12.0
isort==5.13.2